requirements.txt
email_validator
flask
flask-wtf
iso8601
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
requests
gunicorn
icalendar
mailjet_rest
